requirements.txt
#********pip install -r requirements.txt******
django ==4.2  
pillow
django-tinymce